+++ src/CompositePoolNode.h
@@ -13,7 +13,7 @@
 
 #if USE_DELAY_POOLS
 #include "auth/UserRequest.h"
-#include "base/DelayedAsyncCalls.h"
+#include "async/DelayedAsyncCalls.h"
 #include "DelayIdComposite.h"
 #include "DelayPools.h"
 #include "ip/Address.h"
