+++ src/DelayId.cc
@@ -15,7 +15,7 @@
  */
 #if USE_DELAY_POOLS
 #include "acl/FilledChecklist.h"
-#include "base/DelayedAsyncCalls.h"
+#include "async/DelayedAsyncCalls.h"
 #include "client_side_request.h"
 #include "DelayId.h"
 #include "DelayPool.h"
