+++ src/DelayVector.cc
@@ -12,7 +12,7 @@
 
 #if USE_DELAY_POOLS
 #include "async/AsyncCall.h"
-#include "base/DelayedAsyncCalls.h"
+#include "async/DelayedAsyncCalls.h"
 #include "comm/Connection.h"
 #include "DelayVector.h"
