+++ src/Store.h
@@ -9,7 +9,7 @@
 #ifndef SQUID_STORE_H
 #define SQUID_STORE_H
 
-#include "base/DelayedAsyncCalls.h"
+#include "async/DelayedAsyncCalls.h"
 #include "base/Packable.h"
 #include "base/Range.h"
 #include "base/RefCount.h"
