+++ src/http.cc
@@ -16,7 +16,7 @@
 #include "squid.h"
 #include "acl/FilledChecklist.h"
 #include "async/AsyncJobCalls.h"
-#include "base/DelayedAsyncCalls.h"
+#include "async/DelayedAsyncCalls.h"
 #include "base/Raw.h"
 #include "base/TextException.h"
 #include "base64.h"
